requirements.txt
ccxt
tornado==4.5.3
pyyaml
requests
